Break out lower-level request method into its own module
