+++ src/lib/validate.js
@@ -7,10 +7,10 @@ let is = {
   buffer: item => Buffer.isBuffer(item),
   number: item => Number.isInteger(item),
   object: item => item && typeof item === 'object' && !Array.isArray(item),
+  stream: item => item?.on && item?._read && item?._readableState,
   string: item => typeof item === 'string',
   // TODO:
   // - ARN, partial arn
-  // - stream: item => item?.on && item?._read && item?._readableState,
 }
 
 let payloadAliases = [ 'payload', 'body', 'data', 'json' ]
